Fixed drawSpline missing last point issue
Current strategy of drawSpline may miss the last part of curve when the length of points is a even number. 
For example, if the length of points is 14, the code will stop at i = 12 which i + 2 is not < point.length. Therefor, the curve between points[12] and points[13] is missing.

The code added will fix the issue.